+++ R/testit.R
@@ -74,7 +74,7 @@ assert2 = function(fact, exprs, envir, all = TRUE) {
     if (all || (i == n && is.logical(val)) ||
         (length(expr) >= 1 && identical(expr[[1]], as.symbol('(')))) {
       if (all_true(val)) next
-      if (!is.null(fact) && !hide_error()) message('assertion failed: ', fact)
+      if (!is.null(fact)) message('assertion failed: ', fact)
       stop(sprintf(
         ngettext(length(val), '%s is not TRUE', '%s are not all TRUE'),
         deparse_key(expr)
@@ -96,7 +96,7 @@ assert2 = function(fact, exprs, envir, all = TRUE) {
 #' @export
 `%==%` = function(x, y) {
   res = identical(x, y)
-  if (!res && !hide_error() && getOption('testit.asserting', FALSE)) {
+  if (!res && getOption('testit.asserting', FALSE)) {
     mc = match.call()
     info = paste(capture.output({
       cat(deparse_key(mc[[2]]), '(LHS) ==>\n')
@@ -198,7 +198,6 @@ test_pkg = function(package = pkg_name(), dir = c('testit', 'tests/testit'), upd
     withCallingHandlers(
       sys.source2(r, envir = env, top.env = ns),
       error = function(e) {
-        if (hide_error()) return()
         z = if (exists('.traceback', baseenv(), inherits = FALSE)) .traceback(5)
         if (length(z) == 0) return()
         z = z[[1]]
@@ -236,7 +235,6 @@ error_loc = function(x, line = 1, wd = '.') {
 #' @param ... additional arguments passed to \code{\link{grepl}()} for matching
 #'   \code{message} against the condition message (e.g., \code{fixed = TRUE} for
 #'   fixed string matching, or \code{ignore.case = TRUE}).
-#' @param silent logical: should the report of error messages be suppressed?
 #' @return A logical value.
 #' @export
 #' @rdname has_message
@@ -274,11 +272,10 @@ has_warning = function(expr, message = NULL, ...) {
 }
 #' @export
 #' @rdname has_message
-has_error = function(expr, message = NULL, ..., silent = !interactive()) {
+has_error = function(expr, message = NULL, ...) {
   tryCatch({
-    if (silent) silence(expr) else expr; FALSE
+    expr; FALSE
   }, error = function(e) {
-    if (!silent) cat('Error: ', conditionMessage(e), '\n', sep = '')
     match_cond(conditionMessage(e), message, ...)
   })
 }
@@ -288,9 +285,3 @@ match_cond = function(text, message, ...) {
     grepl(message, text, ...)
 }
 
-hide_error = function() getOption('testit.hide.error', FALSE)
-
-silence = function(expr) {
-  op = options(testit.hide.error = TRUE); on.exit(options(op))
-  expr
-}