+++ tests/test-all.R
@@ -1,8 +1,7 @@
 library(testit)
 test_pkg('testit')
 
-op = options(testit.hide.error = TRUE)
 tryCatch(
   test_pkg('testit', 'test-error'),
-  error = function(e) invisible(NULL), finally = function() options(op)
+  error = function(e) invisible(NULL)
 )